Resolves issue #1 when operand is not immediatly read after opcode
The previous implementation assumed that an opcode's operand would always be immediately read following the opcode and before the microprocessor started executing the instruction. This isn't always the case, notably in the JSR instruction, the operand isn't read on the next two clock cycles, the 65c02 will push the return address onto the stack before reading the final byte of the operand.

This change removes any assumptions on when the microprocessor will read the operand, and will only show the decoded instruction once we've seen the full opcode and operand.